requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==21.4.0
bech32==1.2.0
betterproto==2.0.0b4
bip32utils==0.3.post4
bitarray==2.9.2
boltons==21.0.0
certifi==2024.8.30
charset-normalizer==3.3.2
ckzg==2.0.1
cytoolz==0.12.3
ecdsa==0.17.0
eth-account==0.13.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
frozenlist==1.4.1
furl==2.1.3
grpclib==0.4.7
h2==4.1.0
hexbytes==1.2.1
hpack==4.0.0
hyperframe==6.0.1
idna==3.8
MarkupSafe==2.0.1
mnemonic==0.19
multidict==6.1.0
nest-asyncio==1.6.0
orderedmultidict==1.0.1
parsimonious==0.10.0
protobuf==3.20.3
pycryptodome==3.20.0
pydantic==2.9.1
pydantic_core==2.23.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyunormalize==15.1.0
regex==2024.7.24
requests==2.32.3
rlp==4.0.1
six==1.16.0
terra-proto==4.0.2
terra-sdk==4.0.0
toolz==0.12.1
types-requests==2.32.0.20240907
typing_extensions==4.12.2
urllib3==2.2.2
web3==7.2.0
websockets==13.0.1
wrapt==1.16.0
yarl==1.11.1
